Add the Linebender CI script and conform the codebase to it. (#2)
This is based on the latest Linebender CI standard which can be best
seen in the [Xilem
repo](https://github.com/linebender/xilem/tree/main/.github) although
Xilem does have some extra GPU and snapshot testing parts that were
removed here.

The Wasm checks pass so I left it in, although I'm not fully sure if
targeting Wasm is really part of the plan here.

The `.gitignore` file had an entry for `.*` which is too broad. If there
are actual files which still need to be ignored then a more precise
filter can be added in a follow-up PR.

The MSRV in this PR is the same as stable on purpose to keep things
focused on the CI. Finding the actual MSRV and adding the various readme
parts etc related to it are follow-up work.